chore: move FileUtils to a v1/v2 implementation
